requirements.txt
appnope==0.1.4
asttokens==2.4.1
beautifulsoup4==4.9.3
certifi==2024.8.30
charset-normalizer==2.0.12
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dateparser==1.2.0
debugpy==1.8.7
decorator==5.1.1
docutils==0.21.2
echr_extractor==1.0.43
executing==2.1.0
fonttools==4.54.1
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
matplotlib==3.9.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.3
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.26.0
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
statistics==1.0.3.5
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
tzdata==2024.2
tzlocal==5.2
urllib3==1.26.20
wcwidth==0.2.13
